Reorganize wikidata client function
