+++ mira/dkg/client.py
@@ -12,6 +12,7 @@ from typing import TYPE_CHECKING, Any, Dict, Iterable, List, Mapping, Optional, 
 import neo4j.graph
 import networkx
 import pystow
+import requests
 from neo4j import GraphDatabase, Transaction, unit_of_work
 from pydantic import BaseModel, Field
 from tqdm import tqdm
@@ -28,6 +29,10 @@ if TYPE_CHECKING:
 __all__ = ["Neo4jClient", "Entity"]
 
 logger = logging.getLogger(__name__)
+
+#: See documentation for query action at
+#: https://www.wikidata.org/w/api.php?action=help&modules=query
+WIKIDATA_API = "https://www.wikidata.org/w/api.php"
 
 Node: TypeAlias = Mapping[str, Any]
 
@@ -716,6 +721,34 @@ def relation_query(
         range = f"*{min_hops}..{max_hops}"
 
     return rv + range
+
+
+def search_wikidata(text: str) -> List[Entity]:
+    """Search Wikidata with the given text string."""
+    payload = {
+        "action": "wbsearchentities",
+        "search": text,
+        "language": "en",
+        "format": "json",
+        "limit": 50,
+    }
+    res = requests.get(WIKIDATA_API, params=payload)
+    res.raise_for_status()
+    res_json = res.json()
+    results = [_process_result(r) for r in res_json["search"]]
+    # TODO if "search-continue" is available, then there are more results to paginate through.
+    return results
+
+
+def _process_result(record: Mapping[str, Any]) -> Entity:
+    return Entity(
+        id=f"wikidata:{record['id']}",
+        name=record["label"],
+        description=record.get("description", ""),
+        type="class",
+        labels=["wikidata"],
+        obsolete=False,
+    )
 
 
 def _get_search_priority_list():
